Stacking sub header below 300px
The MPV sub header is now set inside an inline-block element with a
min-width of 300px. This causes the dropdown button to stack underneath
(and not above) on smaller devices.

The tabs for Patient Info, Date / Time and Counts in MPV patient
listings have been replaced with a dropdown button for mobile view to
save space.

Also added padding to fluid table layout by adding a container-fluid
div. which means tables don’t run to the edge on all devices.